+++ CHANGELOG.md
@@ -3,10 +3,11 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased][unreleased]
 ### Added
+- Implementation of [Symfony container aware Event Dispatcher](http://symfony.com/doc/current/components/event_dispatcher/container_aware_dispatcher.html).
 - Support for importing resources in config file.
 
 ### Changed
-- Registration of extensions via Symfony Dependency Injection Extension system.
+- Registration of extensions via [Symfony Dependency Injection Extension](http://symfony.com/doc/current/components/dependency_injection/compilation.html) system.
 
 ## 0.1.0 - 2015-04-07
 ### Added